requirements.txt
plotly # for plot representation
pandas
Pillow # for img representation
imageio 